Added 3D admm preliminary code
Preliminary code for ADMM 3D. A main script to setup parameters and
operators, as well as a solver for 3D. This needs to be cleaned up.
Also added script to process zstacks from miniscope, and a script in
python to create Halide compatible zstacks. A function for 3D
convolution via FFT is also included.